Update install.sh
Added a command (and relevant echoes) to attempt to install the missing python packages once before failing the script. Successfully tested on PiZeroW and Pi4, both running the latest 32-bit Bullseye Raspbian. Working through other issues with the 64-bit test device at the moment, but I will confirm this change on that one as well. Feel free to change the echo statement. I tried to keep the style close to what is there.

Also fixed typo in output message of missing package